+++ Zotero/Scenes/Detail/PDF/Models/PDFDocumentAnnotation.swift
@@ -21,7 +21,7 @@ struct PDFDocumentAnnotation {
     let color: String
     let comment: String
     let text: String?
-    var fontSize: CGFloat?
+    var fontSize: UInt?
     var rotation: UInt?
     let sortIndex: String
     let dateModified: Date
